+++ src/utils/chromeDebug.js
@@ -1,5 +1,0 @@
-/* global window */
-export default function chromeDebug(alt) {
-  if (typeof window !== 'undefined') window['alt.js.org'] = alt
-  return alt
-}